simplify the implementation of RAFT, and add utils files to make code concise.
