+++ trinity/common/workflows/envs/alfworld/RAFT_prompt/second_attempt_guidance.j2
@@ -1,0 +1,3 @@
+{{ reward_feedback }}
+
+This is your second attempt. You need to perform better this time. Focus on updating a more thoughtful and comprehensive <experience>...</experience> section that incorporates strategies, common pitfalls, and domain knowledge to help improve task performance and avoid previous mistakes. Learn from your first attempt and apply that knowledge to make better decisions. IMPORTANT: You must update your experience based on what you learned from the first attempt. Also, DO NOT repeat the same action consecutively as this is inefficient and useless - if you find yourself repeating actions, stop and try a different approach. In your responses, do not explicitly mention this is a second attempt or reference previous attempts. Instead, present your actions as if starting fresh, but incorporate general experience and strategies you can summarize for this type of task.
