+++ trinity/common/workflows/envs/alfworld/RAFT_reflect_alfworld_workflow.py
@@ -1,77 +1,25 @@
 # -*- coding: utf-8 -*-
-import json
 import os
-import re
 from datetime import datetime
 from typing import Any, Dict, List, Optional
 
-import torch
-
 from trinity.common.experience import Experience
 from trinity.common.models.model import ModelWrapper
+from trinity.common.workflows.envs.alfworld.RAFT_utils import (
+    create_alfworld_environment,
+    format_observation,
+    generate_default_empty_experience,
+    get_jinja_env,
+    parse_response,
+    process_messages_to_experience,
+    save_task_data,
+    validate_trajectory_format,
+)
 from trinity.common.workflows.workflow import WORKFLOWS, Task, Workflow
 from trinity.utils.log import get_logger
 
 logger = get_logger(__name__)
 logger.setLevel("INFO")
-
-ALFWORLD_SYSTEM_PROMPT = """
-You are an agent interacting with a virtual text-based environment.
-
-## Response Format:
-You MUST use this exact format for every response. All three tags are REQUIRED in sequential order:
-
-<experience>Working principles, strategies, common knowledge, and potential pitfalls relevant to the current task.</experience>\n\n
-<think>your reasoning process</think>\n\n
-<action>exactly one action command</action>
-
-## Action Commands:
-  look:                             look around your current location
-  inventory:                        check your current inventory(you can only have 1 item in your inventory)
-  go to (receptacle):               move to a receptacle
-  open (receptacle):                open a receptacle
-  close (receptacle):               close a receptacle
-  take (object) from (receptacle):  take an object from a receptacle
-  move (object) to (receptacle):    place an object in or on a receptacle
-  examine (something):              examine a receptacle or an object
-  use (object):                     use an object
-  heat (object) with (receptacle):  heat an object using a receptacle
-  clean (object) with (receptacle): clean an object using a receptacle
-  cool (object) with (receptacle):  cool an object using a receptacle
-  slice (object) with (object):     slice an object using a sharp object
-
-For example your output should be like this:
-<experience>In household tasks, I should start by exploring the environment to understand available objects and receptacles. Common pitfalls include forgetting to check inventory capacity and not examining objects before taking action.</experience>\n\n<think> To solve the task, I need first to ... </think>\n\n<action>go to cabinet 1</action>
-
-## Important Note:
-You must ensure that the <experience> section contains descriptions of working principles, strategies, common sense knowledge, and potential pitfalls that are universally applicable to this type of task, rather than generic statements, placeholder content, or overly specific behavioral guidelines.
-"""
-
-
-def format_observation(observation: str):
-    if "Nothing happens." in observation:
-        observation += "Please check if the action you take is valid or you have carefully followed the action format."
-    return "Observation: " + observation
-
-
-def parse_response(response):
-    """Parse all three components from response with a single regex"""
-    try:
-        # Use single regex to extract all three components at once
-        pattern = r"<experience>\s*(.*?)\s*</experience>.*?<think>\s*(.*?)\s*</think>.*?<action>\s*(.*?)\s*</action>"
-        match = re.search(pattern, response, re.DOTALL)
-
-        if match:
-            return {
-                "experience": match.group(1).strip(),
-                "think": match.group(2).strip(),
-                "action": match.group(3).strip(),
-            }
-        else:
-            return {"experience": "", "think": "", "action": ""}
-    except Exception as e:
-        logger.warning(f"Error parsing response: {e}")
-        return {"experience": "", "think": "", "action": ""}
 
 
 @WORKFLOWS.register_module("RAFT_reflect_alfworld_workflow")
@@ -115,6 +63,11 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         os.makedirs(self.sft_dir, exist_ok=True)
         os.makedirs(self.non_sft_dir, exist_ok=True)
 
+        # Setup Jinja2 templates
+        self.jinja_env = get_jinja_env()
+        self.alfworld_system_template = self.jinja_env.get_template("alfworld_system.j2")
+        self.second_attempt_template = self.jinja_env.get_template("second_attempt_guidance.j2")
+
         print(
             f"Initializing RAFTAlfworldWorkflow with RAFT learning, temperature={self.temperature}"
         )
@@ -125,91 +78,9 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         self.game_file_path = task.task_desc or task.raw_task.get("game_file", "")
         self.is_eval = task.is_eval
 
-    def save_task_data(
-        self,
-        task_id: str,
-        first_trajectory: List[Dict[str, str]],
-        first_reward: float,
-        first_steps: int,
-        first_success: bool,
-        second_trajectory: Optional[List[Dict[str, str]]],
-        second_reward: Optional[float],
-        second_steps: Optional[int],
-        second_success: Optional[bool],
-        kept_for_sft: bool,
-        training_data: Optional[List[Dict[str, str]]] = None,
-    ):
-        """Save detailed exploration data to individual task file in appropriate folder"""
-
-        task_data = {
-            "task_id": task_id,
-            "timestamp": datetime.now().isoformat(),
-            "game_file": self.game_file_path,
-            # First exploration
-            "first_exploration": {
-                "trajectory": first_trajectory,
-                "reward": first_reward,
-                "steps": first_steps,
-                "success": first_success,
-            },
-            # Second exploration
-            "second_exploration": {
-                "trajectory": second_trajectory,
-                "reward": second_reward,
-                "steps": second_steps,
-                "success": second_success,
-            },
-            # Training data (clean dialogue format for SFT)
-            "training_data": training_data if training_data is not None else "",
-            "kept_for_sft": kept_for_sft,
-        }
-
-        # Determine folder based on eval mode and SFT data status (following webshop pattern)
-        if self.is_eval:
-            target_dir = self.eval_dir
-        elif kept_for_sft:
-            target_dir = self.sft_dir
-        else:
-            target_dir = self.non_sft_dir
-
-        task_file_path = os.path.join(target_dir, f"{task_id}.json")
-
-        with open(task_file_path, "w", encoding="utf-8") as f:
-            json.dump(task_data, f, ensure_ascii=False, indent=2)
-
-    def validate_response_format(self, parsed: Dict[str, str]) -> bool:
-        """Validate if parsed response has valid content in all required fields"""
-        has_think = len(parsed["think"].strip()) > 0
-        has_experience = len(parsed["experience"].strip()) > 0
-        has_action = len(parsed["action"].strip()) > 0
-
-        return has_think and has_experience and has_action
-
     def create_environment(self, game_file):
         """Create alfworld environment"""
-        try:
-            import textworld
-            import textworld.gym
-            from alfworld.agents.environment.alfred_tw_env import (
-                AlfredDemangler,
-                AlfredExpert,
-                AlfredExpertType,
-            )
-
-            expert = AlfredExpert(expert_type=AlfredExpertType.HANDCODED)
-            request_infos = textworld.EnvInfos(
-                description=True, inventory=True, admissible_commands=True
-            )
-
-            env_id = textworld.gym.register_game(
-                game_file, request_infos, wrappers=[AlfredDemangler(), expert]
-            )
-            env = textworld.gym.make(env_id)
-            return env
-
-        except Exception as e:
-            error_message = f"Error importing AlfworldTWEnv {str(e)}. Please make sure you have installed the alfworld package successfully, following the instructions in https://github.com/alfworld/alfworld"
-            raise ImportError(error_message)
+        return create_alfworld_environment(game_file)
 
     def run_single_rollout(
         self, env
@@ -220,7 +91,7 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         parsed_steps = []  # Store parsed experience, think, action for each step
         action_history = []  # Track last 3 actions for repetition detection
 
-        trajectory.append({"role": "system", "content": ALFWORLD_SYSTEM_PROMPT})
+        trajectory.append({"role": "system", "content": self.alfworld_system_template.render()})
 
         # Track the last reward from environment
         last_reward = 0.0
@@ -242,9 +113,11 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
             # Parse the three components
             parsed = parse_response(response_text)
-            experience_text = parsed["experience"]
-            think_text = parsed["think"]
-            action_text = parsed["action"]
+            experience_text, think_text, action_text = (
+                parsed["experience"],
+                parsed["think"],
+                parsed["action"],
+            )
 
             # Store parsed step for SFT data construction
             parsed_steps.append(
@@ -332,7 +205,7 @@ class RAFTReflectAlfworldWorkflow(Workflow):
     ) -> tuple[List[Dict[str, str]], float, bool, int, List[Dict[str, str]]]:
         """Re-explore with first trajectory as context"""
 
-        env = self.create_environment(self.game_file_path)
+        env = create_alfworld_environment(self.game_file_path)
 
         observation, info = env.reset()
 
@@ -346,12 +219,12 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         context_messages.append(
             {
                 "role": "system",
-                "content": f"{reward_feedback}\n\nThis is your second attempt. You need to perform better this time. Focus on updating a more thoughtful and comprehensive <experience>...</experience> section that incorporates strategies, common pitfalls, and domain knowledge to help improve task performance and avoid previous mistakes. Learn from your first attempt and apply that knowledge to make better decisions. IMPORTANT: You must update your experience based on what you learned from the first attempt. Also, DO NOT repeat the same action consecutively as this is inefficient and useless - if you find yourself repeating actions, stop and try a different approach. In your responses, do not explicitly mention this is a second attempt or reference previous attempts. Instead, present your actions as if starting fresh, but incorporate general experience and strategies you can summarize for this type of task.",
+                "content": self.second_attempt_template.render(reward_feedback=reward_feedback),
             }
         )
 
         # Build clean SFT trajectory (like first trajectory format)
-        sft_trajectory = [{"role": "system", "content": ALFWORLD_SYSTEM_PROMPT}]
+        sft_trajectory = [{"role": "system", "content": self.alfworld_system_template.render()}]
         parsed_steps = []  # Track parsed steps for quality analysis
 
         for step in range(self.max_env_steps):
@@ -374,9 +247,11 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
             # Parse components for quality analysis
             parsed = parse_response(response_text)
-            experience_text = parsed["experience"]
-            think_text = parsed["think"]
-            action_text = parsed["action"]
+            experience_text, think_text, action_text = (
+                parsed["experience"],
+                parsed["think"],
+                parsed["action"],
+            )
 
             parsed_steps.append(
                 {
@@ -401,40 +276,24 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         env.close()
         return sft_trajectory, reward, False, self.max_env_steps, parsed_steps
 
-    def generate_default_empty_experience(
-        self, msg: str = "Unknown error", info=None, metrics=None
-    ) -> Experience:
-        """Generate a default empty experience when errors occur"""
-        if info is None:
-            info = {"error_reason": msg, "rollout_failed": True}
-        if metrics is None:
-            metrics = {"success": 0.0, "reward": 0.0}
-
-        return Experience(
-            tokens=torch.tensor([0], dtype=torch.long),
-            prompt_length=0,
-            action_mask=torch.tensor([False], dtype=torch.bool),
-            info=info,
-            metrics=metrics,
-        )
-
     def eval_alfworld(self) -> List[Experience]:
         """Evaluate a single alfworld trajectory"""
-        env = self.create_environment(self.game_file_path)
+        env = create_alfworld_environment(self.game_file_path)
         try:
             trajectory, reward, done, steps, parsed_steps = self.run_single_rollout(env)
         except Exception as e:
             logger.warning(f"Single rollout failed during eval: {e}")
             env.close()
-            return [self.generate_default_empty_experience(f"Eval rollout failed: {str(e)}")]
+            return [generate_default_empty_experience(f"Eval rollout failed: {str(e)}")]
         env.close()
 
         # Save eval data
         task_id = datetime.now().strftime("%Y%m%d_%H%M%S_%f")
         success = done and reward >= 1
 
-        self.save_task_data(
+        save_task_data(
             task_id=task_id,
+            game_file_path=self.game_file_path,
             first_trajectory=trajectory,
             first_reward=reward,
             first_steps=steps,
@@ -444,11 +303,15 @@ class RAFTReflectAlfworldWorkflow(Workflow):
             second_steps=steps,
             second_success=success,
             kept_for_sft=False,
+            is_eval=self.is_eval,
+            eval_dir=self.eval_dir,
+            sft_dir=self.sft_dir,
+            non_sft_dir=self.non_sft_dir,
             training_data=trajectory,
         )
 
         # Convert trajectory to experience
-        experience = self.generate_default_empty_experience(
+        experience = generate_default_empty_experience(
             msg="Eval completed successfully",
             info={"task_id": task_id, "success": success, "reward": reward, "steps": steps},
             metrics={"success": float(success), "reward": float(reward), "steps": float(steps)},
@@ -458,7 +321,7 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
     def _execute_first_attempt(self, task_id: str) -> tuple:
         """Execute the first attempt and return results"""
-        env = self.create_environment(self.game_file_path)
+        env = create_alfworld_environment(self.game_file_path)
 
         try:
             trajectory, reward, done, steps, parsed_steps = self.run_single_rollout(env)
@@ -469,16 +332,9 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
         env.close()
         success = done and reward >= 1
-        traj_format_valid = self._validate_trajectory_format(parsed_steps)
+        traj_format_valid = validate_trajectory_format(parsed_steps)
 
         return trajectory, reward, done, steps, parsed_steps, success, traj_format_valid
-
-    def _validate_trajectory_format(self, parsed_steps: list) -> bool:
-        """Validate trajectory format"""
-        for step in parsed_steps:
-            if not self.validate_response_format(step):
-                return False
-        return True
 
     def _handle_first_attempt_success(
         self, task_id: str, trajectory: list, reward: float, steps: int, success: bool
@@ -486,8 +342,9 @@ class RAFTReflectAlfworldWorkflow(Workflow):
         """Handle successful first attempt"""
         print("✅ Task completed successfully in the first attempt!")
 
-        self.save_task_data(
+        save_task_data(
             task_id=task_id,
+            game_file_path=self.game_file_path,
             first_trajectory=trajectory,
             first_reward=reward,
             first_steps=steps,
@@ -497,27 +354,24 @@ class RAFTReflectAlfworldWorkflow(Workflow):
             second_steps=None,
             second_success=None,
             kept_for_sft=False,
+            is_eval=self.is_eval,
+            eval_dir=self.eval_dir,
+            sft_dir=self.sft_dir,
+            non_sft_dir=self.non_sft_dir,
             training_data=None,
         )
 
-        try:
-            experience = self.process_messages_to_experience(
-                trajectory, info={"success": success, "reward": reward, "steps": steps}
-            )
-            return [experience]
-        except Exception as e:
-            logger.warning(f"Failed to convert trajectory to experience: {e}")
-            experience = self.generate_default_empty_experience(
-                f"Experience conversion failed: {str(e)}"
-            )
-            return [experience]
+        experience = process_messages_to_experience(
+            self.model, trajectory, info={"success": success, "reward": reward, "steps": steps}
+        )
+        return [experience]
 
     def _handle_invalid_format_success(
         self, success: bool, reward: float, steps: int
     ) -> List[Experience]:
         """Handle case where task succeeded but format is invalid"""
         print("❌ Task completed but trajectory format is invalid, skipping SFT data generation.")
-        experience = self.generate_default_empty_experience(
+        experience = generate_default_empty_experience(
             "Experience conversion failed: Trajectory format invalid",
             metrics={"success": float(success), "reward": float(reward), "steps": float(steps)},
         )
@@ -558,11 +412,7 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
     def _generate_experience_from_sft(self, sft_messages: list, metrics: dict) -> Experience:
         """Generate experience from SFT messages"""
-        try:
-            return self.process_messages_to_experience(sft_messages, info=metrics)
-        except Exception as e:
-            logger.warning(f"Failed to convert messages to experience: {e}")
-            return self.generate_default_empty_experience(f"Experience conversion failed: {str(e)}")
+        return process_messages_to_experience(self.model, sft_messages, info=metrics)
 
     def run(self) -> List[Experience]:
         """Run the RAFT alfworld workflow and return experiences"""
@@ -585,7 +435,7 @@ class RAFTReflectAlfworldWorkflow(Workflow):
                 traj_format_valid,
             ) = self._execute_first_attempt(task_id)
         except Exception as e:
-            return [self.generate_default_empty_experience(f"Training rollout failed: {str(e)}")]
+            return [generate_default_empty_experience(f"Training rollout failed: {str(e)}")]
 
         # Handle first attempt success cases
         if reward >= 1 and traj_format_valid:
@@ -600,14 +450,14 @@ class RAFTReflectAlfworldWorkflow(Workflow):
             trajectory, success, reward, steps
         )
         if error:
-            default_experience = self.generate_default_empty_experience(
+            default_experience = generate_default_empty_experience(
                 f"SFT data construction failed: {str(error)}",
             )
             return [default_experience]
 
         # Validate second attempt and build metrics
         second_success = re_explore_info["new_reward"] >= 1
-        second_traj_format_valid = self._validate_trajectory_format(new_parsed_steps)
+        second_traj_format_valid = validate_trajectory_format(new_parsed_steps)
         metrics = self._build_metrics(reward, steps, new_parsed_steps, re_explore_info)
 
         # Generate experience if conditions are met
@@ -626,7 +476,12 @@ class RAFTReflectAlfworldWorkflow(Workflow):
             )
 
         # Save detailed task data
-        self.save_task_data(
+        save_task_data(
+            game_file_path=self.game_file_path,
+            is_eval=self.is_eval,
+            eval_dir=self.eval_dir,
+            sft_dir=self.sft_dir,
+            non_sft_dir=self.non_sft_dir,
             task_id=task_id,
             first_trajectory=trajectory,
             first_reward=reward,
@@ -642,25 +497,9 @@ class RAFTReflectAlfworldWorkflow(Workflow):
 
         # Return default experience if no valid experience generated
         if not experiences:
-            experiences.append(self.generate_default_empty_experience())
+            experiences.append(generate_default_empty_experience())
 
         return experiences
-
-    def process_messages_to_experience(self, messages, info=None) -> Experience:
-        """Convert messages to experience for training"""
-        if info is None:
-            info = {}
-
-        converted_experience = self.model.convert_messages_to_experience(messages)
-
-        metrics = {}
-        for k, v in info.items():
-            if isinstance(v, float) or isinstance(v, int):
-                metrics[k] = float(v)
-        converted_experience.info = info
-        converted_experience.metrics = metrics
-
-        return converted_experience
 
     def resettable(self) -> bool:
         """Indicate that this workflow can be reset to avoid re-initialization"""
